pyproject.toml
[project]
name = "searchepg"
version = "0.6.0"
description = "Klassifizierung von EPGs (Source MediaServer / DvbViewer)"
authors = [{ name = "Detlef Meyer", email = "dmeyer@dmone.de" }]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "aiohappyeyeballs==2.6.1",
  "aiohttp==3.12.14",
  "aiosignal==1.4.0",
  "annotated-types==0.7.0",
  "anyio==4.9.0",
  "asttokens==3.0.0",
  "attrs==25.3.0",
  "certifi==2025.4.26",
  "charset-normalizer==3.4.2",
  "click==8.2.1",
  "colorama==0.4.6",
  "comm==0.2.2",
  "debugpy==1.8.14",
  "decorator==5.2.1",
  "distro==1.9.0",
  "executing==2.2.0",
  "fastapi==0.115.13",
  "frozenlist==1.7.0",
  "h11==0.16.0",
  "httpcore==1.0.9",
  "httptools==0.6.4",
  "httpx==0.28.1",
  "idna==3.10",
  "ipykernel==6.29.5",
  "ipython==9.3.0",
  "ipython_pygments_lexers==1.1.1",
  "jedi==0.19.2",
  "jiter==0.10.0",
  "jupyter_client==8.6.3",
  "jupyter_core==5.8.1",
  "matplotlib-inline==0.1.7",
  "multidict==6.6.3",
  "nest-asyncio==1.6.0",
  "ollama==0.5.1",
  "openai==1.86.0",
  "packaging==25.0",
  "parso==0.8.4",
  "platformdirs==4.3.8",
  "prompt_toolkit==3.0.51",
  "propcache==0.3.2",
  "psutil==7.0.0",
  "pure_eval==0.2.3",
  "pydantic==2.11.5",
  "pydantic-settings==2.10.0",
  "pydantic_core==2.33.2",
  "Pygments==2.19.1",
  "python-dateutil==2.9.0.post0",
  "python-dotenv==1.1.0",
  "PyYAML==6.0.2",
  "pyzmq==26.4.0",
  "regex==2024.11.6",
  "requests==2.32.4",
  "six==1.17.0",
  "sniffio==1.3.1",
  "stack-data==0.6.3",
  "starlette==0.46.2",
  "tiktoken==0.9.0",
  "tornado==6.5.1",
  "tqdm==4.67.1",
  "traitlets==5.14.3",
  "typing-inspection==0.4.1",
  "typing_extensions==4.14.0",
  "urllib3==2.4.0",
  "uvicorn[standard]==0.34.3",
  "watchfiles==1.1.0",
  "wcwidth==0.2.13",
  "websockets==15.0.1",
  "yarl==1.20.1",
]

[tool.setuptools]
packages = ["backend"]


[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

